app/requirements.txt
python-dotenv==1.0.0
gspread
google-auth
redis